% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{call_site_key}
\alias{call_site_key}
\title{Create a string key for a call site (file + line)}
\usage{
call_site_key(file, line)
}
\arguments{
\item{file}{Character basename or full path}

\item{line}{Integer line number}
}
\value{
Character string key
}
\description{
Used as named-list / environment key instead of a hashable object.
}
\keyword{internal}
