% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{clean_dataset_name}
\alias{clean_dataset_name}
\title{Clean a SAS dataset name}
\usage{
clean_dataset_name(name)
}
\arguments{
\item{name}{Character raw dataset token to clean}
}
\value{
Character cleaned, lower-cased dataset name, or \code{NULL} when empty
}
\description{
Clean a SAS dataset name
}
