% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{escape_txt_value}
\alias{escape_txt_value}
\title{Escape a value for txt format (quote if contains spaces or equals)}
\usage{
escape_txt_value(value)
}
\arguments{
\item{value}{Character value}
}
\value{
Character possibly quoted
}
\description{
Escape a value for txt format (quote if contains spaces or equals)
}
\keyword{internal}
