% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{extract_macro_name_from_def}
\alias{extract_macro_name_from_def}
\title{Get the macro name from a macro definition}
\usage{
extract_macro_name_from_def(macro_def)
}
\arguments{
\item{macro_def}{Macro definition record}
}
\value{
Character macro name
}
\description{
Get the macro name from a macro definition
}
