% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{find_statement_end_semicolon}
\alias{find_statement_end_semicolon}
\title{Find first semicolon not inside parentheses or quotes}
\usage{
find_statement_end_semicolon(s)
}
\arguments{
\item{s}{Character string to scan}
}
\value{
Integer 1-based index of the terminating semicolon, or -1 if none
}
\description{
Find first semicolon not inside parentheses or quotes
}
