% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{get_operation_style}
\alias{get_operation_style}
\title{Get the DOT style string for an operation node}
\usage{
get_operation_style(op_node, target_datasets)
}
\arguments{
\item{op_node}{Named list (OperationNode)}

\item{target_datasets}{Character vector of target dataset names}
}
\value{
Character DOT style attribute string
}
\description{
Get the DOT style string for an operation node
}
\keyword{internal}
