% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{is_input_lib}
\alias{is_input_lib}
\title{Check if a library name is an external source library}
\usage{
is_input_lib(lib)
}
\arguments{
\item{lib}{Character library name}
}
\value{
Logical
}
\description{
Check if a library name is an external source library
}
