% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{make_node_id}
\alias{make_node_id}
\title{Create a valid Graphviz node ID from an operation node}
\usage{
make_node_id(op_node, display_file)
}
\arguments{
\item{op_node}{Named list (OperationNode)}

\item{display_file}{Character display path (relative to sas_dir)}
}
\value{
Character valid Graphviz ID
}
\description{
Sanitizes the combination of dataset + display_file + line_number.
}
\keyword{internal}
