% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{new_call_stack_frame}
\alias{new_call_stack_frame}
\title{Create a new CallStackFrame (named list)}
\usage{
new_call_stack_frame(file_path, line, frame_type, macro_name = NULL)
}
\arguments{
\item{file_path}{Character full path}

\item{line}{Integer resume line after include/macro call}

\item{frame_type}{Character: "include" or "macro"}

\item{macro_name}{Optional character: macro name if frame_type == "macro"}
}
\value{
Named list
}
\description{
Create a new CallStackFrame (named list)
}
\keyword{internal}
