% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph_nodes.R
\name{new_og_macro_definition}
\alias{new_og_macro_definition}
\title{Create a new MacroDefinition (named list)}
\usage{
new_og_macro_definition(
  name,
  file_path,
  start_line,
  end_line,
  params = character(0)
)
}
\arguments{
\item{name}{Character macro name (lower-cased)}

\item{file_path}{Character full path to where macro is defined}

\item{start_line}{Integer line with the \verb{\%macro} statement (1-based)}

\item{end_line}{Integer line with the \verb{\%mend} statement (1-based)}

\item{params}{Character vector of parameter names}
}
\value{
Named list
}
\description{
Create a new MacroDefinition (named list)
}
\keyword{internal}
