% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operations_graph.R
\name{OperationsGraphGenerator}
\alias{OperationsGraphGenerator}
\title{Operations Graph Generator R6 Class}
\description{
Operations Graph Generator R6 Class

Operations Graph Generator R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OperationsGraphGenerator-new}{\code{OperationsGraphGenerator$new()}}
\item \href{#method-OperationsGraphGenerator-log}{\code{OperationsGraphGenerator$log()}}
\item \href{#method-OperationsGraphGenerator-load_manifests}{\code{OperationsGraphGenerator$load_manifests()}}
\item \href{#method-OperationsGraphGenerator-build_filename_alias_map}{\code{OperationsGraphGenerator$build_filename_alias_map()}}
\item \href{#method-OperationsGraphGenerator-build_macro_map}{\code{OperationsGraphGenerator$build_macro_map()}}
\item \href{#method-OperationsGraphGenerator-walk_code}{\code{OperationsGraphGenerator$walk_code()}}
\item \href{#method-OperationsGraphGenerator-generate_dot}{\code{OperationsGraphGenerator$generate_dot()}}
\item \href{#method-OperationsGraphGenerator-generate_txt}{\code{OperationsGraphGenerator$generate_txt()}}
\item \href{#method-OperationsGraphGenerator-generate_graph_md}{\code{OperationsGraphGenerator$generate_graph_md()}}
\item \href{#method-OperationsGraphGenerator-generate_code_extracts_md}{\code{OperationsGraphGenerator$generate_code_extracts_md()}}
\item \href{#method-OperationsGraphGenerator-generate_spec_index_md}{\code{OperationsGraphGenerator$generate_spec_index_md()}}
\item \href{#method-OperationsGraphGenerator-generate_spec_index_json}{\code{OperationsGraphGenerator$generate_spec_index_json()}}
\item \href{#method-OperationsGraphGenerator-clone}{\code{OperationsGraphGenerator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-new"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$new(sas_dir, entrypoint, manifest_paths)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-log"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-log}{}}}
\subsection{Method \code{log()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$log(msg)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-load_manifests"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-load_manifests}{}}}
\subsection{Method \code{load_manifests()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$load_manifests()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-build_filename_alias_map"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-build_filename_alias_map}{}}}
\subsection{Method \code{build_filename_alias_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$build_filename_alias_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-build_macro_map"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-build_macro_map}{}}}
\subsection{Method \code{build_macro_map()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$build_macro_map()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-walk_code"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-walk_code}{}}}
\subsection{Method \code{walk_code()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$walk_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_dot"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_dot}{}}}
\subsection{Method \code{generate_dot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_dot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_txt"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_txt}{}}}
\subsection{Method \code{generate_txt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_txt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_graph_md"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_graph_md}{}}}
\subsection{Method \code{generate_graph_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_graph_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_code_extracts_md"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_code_extracts_md}{}}}
\subsection{Method \code{generate_code_extracts_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_code_extracts_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_spec_index_md"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_spec_index_md}{}}}
\subsection{Method \code{generate_spec_index_md()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_spec_index_md()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-generate_spec_index_json"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-generate_spec_index_json}{}}}
\subsection{Method \code{generate_spec_index_json()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$generate_spec_index_json()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OperationsGraphGenerator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OperationsGraphGenerator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OperationsGraphGenerator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
