% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_step.R
\name{parse_data_output_datasets}
\alias{parse_data_output_datasets}
\title{Parse output dataset names from DATA statement}
\usage{
parse_data_output_datasets(data_statement)
}
\arguments{
\item{data_statement}{Character full DATA statement}
}
\value{
Character vector of output dataset names, or \code{NULL} when none
}
\description{
Parse output dataset names from DATA statement
}
