% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename.R
\name{parse_filename_statements}
\alias{parse_filename_statements}
\title{Parse FILENAME statements to map filerefs to actual paths}
\usage{
parse_filename_statements(filepath)
}
\arguments{
\item{filepath}{Character path to the SAS source file}
}
\value{
List with \code{filename_refs} and \code{file_fileref_definitions}
}
\description{
Parse FILENAME statements to map filerefs to actual paths
}
