% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/macro.R
\name{parse_macro_definitions}
\alias{parse_macro_definitions}
\title{Parse macro definitions from a SAS file}
\usage{
parse_macro_definitions(filepath)
}
\arguments{
\item{filepath}{Character path to the SAS source file}
}
\value{
List of macro definition records keyed by macro name
}
\description{
Parse macro definitions from a SAS file
}
