% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_sort.R
\name{parse_proc_sort}
\alias{parse_proc_sort}
\title{Parse PROC SORT}
\usage{
parse_proc_sort(lines, start_idx, filepath)
}
\arguments{
\item{lines}{Character vector of file lines}

\item{start_idx}{1-based index}

\item{filepath}{File path string}
}
\value{
An operation (named list) or NULL
}
\description{
Parse PROC SORT
}
