% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{replace_operation}
\alias{replace_operation}
\title{Create a copy of an operation with some fields replaced}
\usage{
replace_operation(op, ...)
}
\arguments{
\item{op}{Named list operation to copy}

\item{...}{Named fields to replace in the copy}
}
\value{
Named list with the replaced fields and coerced integer line fields
}
\description{
Create a copy of an operation with some fields replaced
}
