% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzer.R
\name{SASLineageAnalyzer}
\alias{SASLineageAnalyzer}
\title{SAS Lineage Analyzer R6 Class}
\description{
SAS Lineage Analyzer R6 Class

SAS Lineage Analyzer R6 Class
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SASLineageAnalyzer-new}{\code{SASLineageAnalyzer$new()}}
\item \href{#method-SASLineageAnalyzer-parse_all_sas_files}{\code{SASLineageAnalyzer$parse_all_sas_files()}}
\item \href{#method-SASLineageAnalyzer-parse_sas_file}{\code{SASLineageAnalyzer$parse_sas_file()}}
\item \href{#method-SASLineageAnalyzer-trace_dependencies}{\code{SASLineageAnalyzer$trace_dependencies()}}
\item \href{#method-SASLineageAnalyzer-generate_json_manifest}{\code{SASLineageAnalyzer$generate_json_manifest()}}
\item \href{#method-SASLineageAnalyzer-generate_report}{\code{SASLineageAnalyzer$generate_report()}}
\item \href{#method-SASLineageAnalyzer-deduplicate_operations}{\code{SASLineageAnalyzer$deduplicate_operations()}}
\item \href{#method-SASLineageAnalyzer-clone}{\code{SASLineageAnalyzer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-new"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$new(sas_dir, include_search_roots = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-parse_all_sas_files"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-parse_all_sas_files}{}}}
\subsection{Method \code{parse_all_sas_files()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$parse_all_sas_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-parse_sas_file"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-parse_sas_file}{}}}
\subsection{Method \code{parse_sas_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$parse_sas_file(filepath)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-trace_dependencies"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-trace_dependencies}{}}}
\subsection{Method \code{trace_dependencies()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$trace_dependencies(target_dataset, max_depth = 20L)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-generate_json_manifest"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-generate_json_manifest}{}}}
\subsection{Method \code{generate_json_manifest()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$generate_json_manifest(
  target_dataset,
  output_file = "lineage-manifest.json",
  filter_orphan_nodes = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$generate_report(
  target_dataset,
  output_file = "lineage-report.md",
  filter_orphan_nodes = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-deduplicate_operations"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-deduplicate_operations}{}}}
\subsection{Method \code{deduplicate_operations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$deduplicate_operations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SASLineageAnalyzer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SASLineageAnalyzer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SASLineageAnalyzer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
