% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saslineager-package.R
\docType{package}
\name{saslineager-package}
\alias{saslineager}
\alias{saslineager-package}
\title{saslineager: SAS Data Lineage Analyzer}
\description{
Traces SAS dataset lineage across source files, producing JSON manifests and markdown reports of the dependency graph.
}
\author{
\strong{Maintainer}: Matthieu de Canteloube \email{matthieu.de-canteloube@theodo.com}

}
\keyword{internal}
