% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proc_format.R
\name{scan_format_refs}
\alias{scan_format_refs}
\title{Return ordered, de-duplicated user-defined format names referenced via \code{put()}}
\usage{
scan_format_refs(code_lines)
}
\arguments{
\item{code_lines}{Character vector of source code lines to scan}
}
\value{
Character vector of referenced user-defined format names
}
\description{
Return ordered, de-duplicated user-defined format names referenced via \code{put()}
}
