% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{scan_macro_var_refs}
\alias{scan_macro_var_refs}
\title{Return ordered, de-duplicated macro variable names referenced in code}
\usage{
scan_macro_var_refs(code_lines)
}
\arguments{
\item{code_lines}{Character vector of source code lines to scan}
}
\value{
Character vector of referenced macro variable names
}
\description{
Return ordered, de-duplicated macro variable names referenced in code
}
