% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{strip_macro_control_flow}
\alias{strip_macro_control_flow}
\title{Replace macro control-flow tokens with spaces}
\usage{
strip_macro_control_flow(s)
}
\arguments{
\item{s}{Character string possibly containing macro control-flow tokens}
}
\value{
Character string with control-flow tokens replaced by spaces
}
\description{
Replace macro control-flow tokens with spaces
}
