% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{strip_sas_block_comments}
\alias{strip_sas_block_comments}
\title{Remove SAS block comments from a string}
\usage{
strip_sas_block_comments(s)
}
\arguments{
\item{s}{Character string possibly containing \verb{/* ... */} comments}
}
\value{
Character string with block comments replaced by spaces
}
\description{
Remove SAS block comments from a string
}
