% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includes.R
\name{substitute_macro_vars}
\alias{substitute_macro_vars}
\title{Substitute known macro-var references in a path}
\usage{
substitute_macro_vars(path, macro_variables)
}
\arguments{
\item{path}{Character path possibly containing macro variable references}

\item{macro_variables}{Named list of known macro variable values}
}
\value{
Character path with known macro variables substituted
}
\description{
Substitute known macro-var references in a path
}
